The error causing an invalid query exception in the database has been fixed.

+++ App/Repositories/TeamMembersRepository.cs
@@ -9,36 +9,41 @@ namespace ConstructionManagementApp.App.Repositories
     internal class TeamMembersRepository
     {
         private readonly AppDbContext _context;
+        private readonly UserRepository _userRepository;
 
-        public TeamMembersRepository(AppDbContext context)
+        public TeamMembersRepository(AppDbContext context, UserRepository userRepository)
         {
             _context = context;
+            _userRepository = userRepository;
         }
 
-        public void AddMemberToTeam(int teamId, int userId)
+        public void AddMemberToTeam(int teamId, string userName)
         {
             var team = _context.Teams.FirstOrDefault(t => t.Id == teamId);
             if (team == null)
                 throw new KeyNotFoundException($"Nie znaleziono zespołu o ID {teamId}.");
 
-            var user = _context.Users.FirstOrDefault(u => u.Id == userId);
+            var user = _userRepository.GetUserByUsername(userName);
             if (user == null)
-                throw new KeyNotFoundException($"Nie znaleziono użytkownika o ID {userId}.");
+                throw new KeyNotFoundException($"Nie znaleziono użytkownika o nazwie {userName}.");
 
             // Sprawdzenie, czy użytkownik już jest członkiem zespołu
-            if (_context.TeamMembers.Any(tm => tm.TeamId == teamId && tm.UserId == userId))
-                throw new InvalidOperationException($"Użytkownik o ID {userId} jest już członkiem zespołu o ID {teamId}.");
+            if (_context.TeamMembers.Any(tm => tm.TeamId == teamId && tm.UserId == user.Id))
+                throw new InvalidOperationException($"Użytkownik {userName} jest już członkiem zespołu o ID {teamId}.");
 
-            var teamMember = new TeamMembers(teamId, userId);
+            var teamMember = new TeamMembers(teamId, user.Id);
             _context.TeamMembers.Add(teamMember);
             _context.SaveChanges();
         }
 
-        public void RemoveMemberFromTeam(int teamId, int userId)
+        public void RemoveMemberFromTeam(int teamId, string userName)
         {
-            var teamMember = _context.TeamMembers.FirstOrDefault(tm => tm.TeamId == teamId && tm.UserId == userId);
+            var user = _userRepository.GetUserByUsername(userName);
+            if (user == null)
+                throw new KeyNotFoundException($"Nie znaleziono użytkownika o nazwie {userName}."); 
+            var teamMember = _context.TeamMembers.FirstOrDefault(tm => tm.TeamId == teamId && tm.UserId == user.Id);
             if (teamMember == null)
-                throw new KeyNotFoundException($"Nie znaleziono użytkownika o ID {userId} w zespole o ID {teamId}.");
+                throw new KeyNotFoundException($"Nie znaleziono użytkownika {userName} w zespole o ID {teamId}.");
 
             _context.TeamMembers.Remove(teamMember);
             _context.SaveChanges();
